extract common testutil.js for tests, rename single-worker project
