+++ test-projects/single-worker/tests/simple/file.jpg
