+++ test-projects/single-worker/tests/simple/rendition.jpg
